fix: aislar serializer de cursos pwa
