+++ pwa/api_serializers.py
@@ -6,7 +6,7 @@ from django.db.utils import OperationalError, ProgrammingError
 from rest_framework import serializers
 
 from comunicados.models import Comunicado, ComunicadoAdjunto, SubtipoComunicado
-from comedores.models import CursoAppMobile, Nomina
+from comedores.models import Nomina
 from comedores.models import ActividadColaboradorEspacio, ColaboradorEspacio
 from core.models import Dia, Sexo
 
@@ -960,31 +960,6 @@ class MensajeEspacioPWASerializer(serializers.ModelSerializer):
         if obj.subtipo == SubtipoComunicado.INSTITUCIONAL:
             return "general"
         return "espacio"
-
-
-class CursoAppMobilePWASerializer(serializers.ModelSerializer):
-    imagen_url = serializers.SerializerMethodField()
-
-    class Meta:
-        model = CursoAppMobile
-        fields = (
-            "id",
-            "nombre",
-            "link",
-            "descripcion",
-            "programa_objetivo",
-            "es_recomendado",
-            "activo",
-            "orden",
-            "imagen_url",
-        )
-
-    def get_imagen_url(self, obj):
-        if not obj.imagen:
-            return None
-        request = self.context.get("request")
-        url = obj.imagen.url
-        return request.build_absolute_uri(url) if request else url
 
 
 class PushSubscriptionPWAConfigSerializer(serializers.Serializer):
